% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_doc.R
\name{add_doc}
\alias{add_doc}
\title{Add meta document to database}
\usage{
add_doc(type, ref, doc, conn, overwrite = FALSE)
}
\arguments{
\item{type}{character string with type, freely definable.}

\item{ref}{character string with reference, freely definable.}

\item{doc}{R object to store}

\item{conn}{database connection object

It is up to the user to organise the meta-data in the database.
Meta-data objects are ordinary R objects like character strings,
lists or data.frames. These objects can be organised using the type
and ref fields. For example, a data.frame with specifications of a
certain sensor at a stations, can be stored using 'sensor' as type
and the station id as ref (reference).}
}
\description{
Adds a meta-data document to the database. This meta-data document
is an R object (e.g. list or data.frame) which is stored in the
database. Meta_documents are findable using their type and
reference.
}
