% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_database_tables.R
\name{create_database_tables}
\alias{create_database_tables}
\title{Create new set of database tables}
\usage{
create_database_tables(conn)
}
\arguments{
\item{pool}{pool object with database connection}
}
\description{
This function creates the database tables for this package. It uses
a pool object with the database connection
}
