% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_to_matrix.R
\name{datetime_to_matrix}
\alias{datetime_to_matrix}
\title{Helper function to converts datetime strings into matrix}
\usage{
datetime_to_matrix(Tstart, Tend)
}
\arguments{
\item{Tstart}{Start time}

\item{Tend}{end time}
}
\value{
A single row matrix suitable for the download handler

The start and end time in this function must be given in a string
which can be interpreted by lubridate's \code{as_datetime} function.
These strings are converted into timestamps in seconds (POSIXct
timestamp) and returned as matrix. The first column is the start
time, the second column is the end time.
}
\description{
The download handler functions require a date-time range as input.
This is a matrix with timestamps in seconds. This function converts
an ordinary time range, as string, into such matrix.
}
