% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_exists.R
\name{doc_exists}
\alias{doc_exists}
\title{Checks if a meta-document exists}
\usage{
doc_exists(type, ref, conn)
}
\arguments{
\item{type}{character string with type}

\item{ref}{character string with reference}

\item{conn}{database connection object}
}
\value{
TRUE if document exists, FALSE otherwise

A meta document is identified by a unique type and ref combination.
This function can be used to check if a document with given type
and ref already exists in the meta table.
}
\description{
This functions checks if a meta document allready exists. A
document must have an unique type and ref combination and
duplicated combinations are not allowed
}
