% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Downloads data using download handler}
\usage{
download_data(station, Tstart, Tend, fun, conn)
}
\arguments{
\item{station}{station id}

\item{Tstart}{start of time range}

\item{Tend}{end of time range}

\item{fun}{name of function to use as downloadhandler}

\item{conn}{database connection object}
}
\value{
a data.frame with the downloaded measurements

This function does three things. It first determines which (parts
of the) time ranges are missing from the database. Next, it will
download the missing data. Third, it adds the data to the database
and updates the table wich the available time ranges

This function uses a seperate function as download handler.
Depending on the source (file, database or API) a specific function
must be available

TODO: complete this, but first explain working in vignette, add
summery here
}
\description{
This function downloads measurements for a specific time range and
adds these to the database. Depending on the source (e.g. an API)
of the data, a download handler is needed. This download handler is
a function wich get's the data for the time range and returns a
data.frame with using a specific format.
}
