% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data_fun}
\alias{download_data_fun}
\title{Example download handler function}
\usage{
download_data_fun(x, station, conn)
}
\description{
This is an example of a download function. We need to explain this
further.
}
