% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_database_tables.R
\name{drop_database_tables}
\alias{drop_database_tables}
\title{Drop data base tables}
\usage{
drop_database_tables(conn)
}
\arguments{
\item{conn}{pool connection object

This function drops all tables from the connected database. Please
use this with caution, because this operation can not be undone}
}
\description{
This function drops (removes) all database tables. Use with
caution.
}
