% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doc.R
\name{get_doc}
\alias{get_doc}
\title{Gets a meta-document from the database}
\usage{
get_doc(type, ref, conn)
}
\arguments{
\item{type}{character string with type}

\item{ref}{character string with reference}

\item{conn}{database connection object}
}
\value{
The R object or NA if object is not found

A meta-data document is an R object which are stored in the
database. Each object must have a type and reference (ref) so they
are findable

This functions gets the specific data object from the database and
returns it as an R object (the same object when it was stored). If
the object is not found, this function returns NA
}
\description{
This function gets a meta-data document from the database, using
type and reference as search criteria.
}
