% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_station_info.R
\name{insert_sensor_info}
\alias{insert_sensor_info}
\title{Insert sensor information}
\usage{
insert_sensor_info(station, lat, lon, conn, timestamp = lubridate::now())
}
\arguments{
\item{station}{station identifier}

\item{lat}{latitude coordinate}

\item{lon}{longitude coordinate}

\item{conn}{database connection object

This function insers station information into the database 'sensor'
table. Each record contains a station identifier, lat/lon
coordinates and a timestamp. The timestamp can be used to track
mobile sensors. If no timestamp is given, the current date and time
is added.}

\item{timestamp}{timestamp, default to current time}
}
\description{
Insert sensor information (station identifier and lat / lon
coordinates) into the database.
}
