% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_doc.R
\name{list_doc}
\alias{list_doc}
\title{Lists all references for a meta-document type}
\usage{
list_doc(type, conn)
}
\arguments{
\item{type}{character string with type}

\item{conn}{database connection object}
}
\value{
character vector of references (ref) for the given type, or
NA if no documents of that type exist

A meta-data document is an R object which is stored in the database.
Each object has a type and reference (ref) so they are findable.
This function returns all references belonging to a certain type.
}
\description{
This function lists the references (ref) of all meta-data documents
in the database for a given type.
}
